refactor(ui-react): moving modules around
