--- packages/ui-react/src/modules/App/index.js
@@ -1,1 +1,0 @@
-export { MODE } from './constants'
