--- packages/ui-react/src/modules/App/mode.selectors.js
@@ -1,2 +1,0 @@
-
-export const selectMode = (state) => (state.mode)
