--- packages/ui-react/src/modules/App/selectedNodeId.selector.js
@@ -1,2 +1,0 @@
-
-export const selectSelectedNodeId = (state) => (state.setSelectedNodeId)
