--- packages/ui-react/src/modules/Chip/ChipDetail/index.js
@@ -1,3 +1,0 @@
-import Details from './Details.container'
-
-export default Details
