+++ packages/ui-react/src/modules/Chip/components.js
@@ -1,1 +1,2 @@
 export { ChipList } from './ChipList/ChipList.container'
+export { ChipDetail } from './ChipDetail/ChipDetail.container'
