+++ packages/ui-react/src/modules/CircuitBoard/CircuitBoard.jsx
@@ -40,7 +40,7 @@ import * as d3 from 'd3'
 
 import styles from './CircuitBoard.module.css'
 import { SVG_HEIGHT, SVG_WIDTH, NODE_TYPE } from './constants'
-import { MODE } from '../App'
+import { MODE } from '../Mode'
 
 export function CircuitBoard({ mode, setMode, displaySettings }) {
