+++ packages/ui-react/src/modules/CircuitBoard/selectedNodeId.selector.js
@@ -1,0 +1,3 @@
+import { prop } from 'ramda'
+
+export const selectedNodeIdSelector = prop ('selectedNodeId')
