+++ packages/ui-react/src/modules/CircuitBoard/selectors.js
@@ -1,7 +1,7 @@
 import { createSelector } from '@reduxjs/toolkit'
 import { prop } from 'ramda'
 
-import { selectMode } from '../App/mode.selectors'
+import { modeSelector } from '../Mode/mode.selectors'
 import { selectDisplaySettings } from '../App/displaySettings.selectors'
 
 export const selectCircuitBoard = (state) => (state.circuitBoard)
@@ -11,7 +11,7 @@ export const selectOutputs = createSelector(selectCircuitBoard, prop('outputs'))
 export const selectNodes = createSelector(selectCircuitBoard, prop('nodes'))
 
 export const selector = createSelector(
-  [selectMode, selectDisplaySettings],
+  [modeSelector, selectDisplaySettings],
   (mode, displaySettings) => ({
     mode,
     displaySettings,
