+++ packages/ui-react/src/modules/Mode/index.js
@@ -1,0 +1,2 @@
+export { MODE } from './mode.constants'
+export { reducer as mode } from './mode.reducer'
