+++ packages/ui-react/src/modules/Mode/mode.selectors.js
@@ -1,0 +1,2 @@
+
+export const modeSelector = (state) => (state.mode)
